Had to modify this test. Works better if results/ does not exist

+++ pyinseq/tests/CLEAN_UP.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 # Remove the past pyinseq test directroy with cwd as context
-rm -r results/
+# shellcheck disable=SC1049
+if [ -d "results/" ]; then
+  rm -r results/
+fi
